feat: create auth flow for general users
